Fix making app-update-yml file for mac
